Fix bug

+++ gc.c
@@ -88,7 +88,7 @@ object gc_new_lambda(object env, object params, object body)
   return new_lambda(env, params, body, FALSE);
 }
 
-object gc_new_bytes(int val)
+object gc_new_bytes(int64_t val)
 {
   object o;
   o = gc_alloc(sizeof(struct xint));
@@ -98,7 +98,7 @@ object gc_new_bytes(int val)
   return o;
 }
 
-object gc_new_xint(int val)
+object gc_new_xint(int64_t val)
 {
   if (byte_range_p(val)) return object_bytes[val];
   return gc_new_bytes(val);
